backend/requirements.txt
fastapi
motor
pydantic-settings
uvicorn
python-dotenv
python-multipart
google-generativeai
pillow
elevenlabs
fastapi
motor
pydantic-settings
uvicorn
python-dotenv
python-multipart
google-generativeai
pillow
elevenlabs
twilio
websockets
ngrok
pytest
analytics.stats_engine
